bentoml/tests/requirements.txt
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-timeout>=2.1.0
requests>=2.31.0
pillow>=9.0.0
numpy>=1.20.0

